# core/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/core/api/v2/set-record": {
      "put": {
        "tags": [
          "set-record-controller"
        ],
        "operationId": "updateSetRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSetRecord"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceSetRecord"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "set-record-controller"
        ],
        "operationId": "createSetRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSetRecord"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceSetRecord"
                }
              }
            }
          }
        }
      }
    },
...
    "/core/api/v2/set-record/{id}": {
      "get": {
        "tags": [
          "set-record-controller"
        ],
        "operationId": "readSetRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceSetRecord"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "set-record-controller"
        ],
        "operationId": "deleteSetRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/core/api/v2/records-view": {
      "get": {
        "tags": [
          "record-view-controller"
        ],
        "operationId": "getAllRecords",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ResourceRecordViewFilter"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelEntityModelResourceRecordView"
                }
              }
            }
          }
        }
      }
    },
    "/core/api/v2/records-view/{id}": {
      "get": {
        "tags": [
          "record-view-controller"
        ],
        "operationId": "getRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceRecordView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "time": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpdateSetRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "default": "00000000-0000-0000-0000-000000000000"
          },
          "title": {
            "type": "string"
          },
          "aHref": {
            "type": "string"
          },
          "texts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "postAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshAt": {
            "type": "string",
            "format": "date-time"
          },
          "visible": {
            "type": "boolean"
          },
          "flags": {
            "type": "integer",
            "format": "int32"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "id",
          "parentId",
          "refreshAt"
        ]
      },
      "ResourceSetRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "visible": {
            "type": "boolean"
          },
          "flags": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "aHref": {
            "type": "string"
          },
          "texts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "postAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshAt": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        }
      },
...
      "NewSetRecord": {
        "type": "object",
        "properties": {
          "visible": {
            "type": "boolean"
          },
          "flags": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "default": "00000000-0000-0000-0000-000000000000"
          },
          "title": {
            "type": "string"
          },
          "aHref": {
            "type": "string"
          },
          "texts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "postAt": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "parentId",
          "postAt",
          "texts"
        ]
      },
...
      "ResourceRecordViewFilter": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "fromTime": {
            "type": "string",
            "format": "date-time"
          },
          "fromDate": {
            "type": "string",
            "format": "date-time"
          },
          "toDate": {
            "type": "string",
            "format": "date-time"
          },
          "withDisabled": {
            "type": "boolean"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EntityModelResourceRecordView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "visible": {
            "type": "boolean"
          },
          "flags": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "Base",
              "Blob",
              "Json",
              "Set",
              "Text",
              "Vector",
              "Xml"
            ]
          },
          "userName": {
            "type": "string"
          },
          "postAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastChangedTime": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "aHref": {
            "type": "string"
          },
          "blob": {
            "type": "string",
            "format": "byte"
          },
          "json": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "texts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "fileName": {
            "type": "string"
          },
          "html": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "markdown": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "vector": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "xml": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelEntityModelResourceRecordView": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "resourceRecordViewList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelResourceRecordView"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
      "ResourceRecordView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "visible": {
            "type": "boolean"
          },
          "flags": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "Base",
              "Blob",
              "Json",
              "Set",
              "Text",
              "Vector",
              "Xml"
            ]
          },
          "userName": {
            "type": "string"
          },
          "postAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastChangedTime": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "aHref": {
            "type": "string"
          },
          "blob": {
            "type": "string",
            "format": "byte"
          },
          "json": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "texts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "fileName": {
            "type": "string"
          },
          "html": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "markdown": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "vector": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "xml": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    }
  }
}